share workflow content getting code
