+++ pkg/cmd/workflow/view/http.go
@@ -1,9 +1,7 @@
 package view
 
 import (
-	"encoding/base64"
 	"fmt"
-	"net/url"
 
 	"github.com/cli/cli/api"
 	"github.com/cli/cli/internal/ghrepo"
@@ -14,31 +12,6 @@ import (
 type workflowRuns struct {
 	Total int
 	Runs  []runShared.Run
-}
-
-func getWorkflowContent(client *api.Client, repo ghrepo.Interface, ref string, workflow *shared.Workflow) (string, error) {
-	path := fmt.Sprintf("repos/%s/contents/%s", ghrepo.FullName(repo), workflow.Path)
-	if ref != "" {
-		q := fmt.Sprintf("?ref=%s", url.QueryEscape(ref))
-		path = path + q
-	}
-
-	type Result struct {
-		Content string
-	}
-
-	var result Result
-	err := client.REST(repo.RepoHost(), "GET", path, nil, &result)
-	if err != nil {
-		return "", err
-	}
-
-	decoded, err := base64.StdEncoding.DecodeString(result.Content)
-	if err != nil {
-		return "", fmt.Errorf("failed to decode workflow file: %w", err)
-	}
-
-	return string(decoded), nil
 }
 
 func getWorkflowRuns(client *api.Client, repo ghrepo.Interface, workflow *shared.Workflow) (workflowRuns, error) {
